+++ lib/shared/services/network_const.dart
@@ -1,11 +1,11 @@
-class ApiConsts {
+class NetworkConsts {
   /// the root url for the api: https://api.ourjourneys.lukecreated.com
   static const String apiRootUrl = 'https://api.ourjourneys.lukecreated.com';
 
-  /// the version of the api: 1
-  static const int apiVersion = 1;
+  /// the version of the api: 2
+  static const int apiVersion = 2;
 
-  /// the base url for the api: https://api.ourjourneys.lukecreated.com/v1
+  /// the base url for the api: https://api.ourjourneys.lukecreated.com/v2
   static const String apiBaseUrl = '$apiRootUrl/v$apiVersion';
 
   /// the header for the api: Content-Type
@@ -19,4 +19,7 @@ class ApiConsts {
 
   /// the header for the api: Bearer
   static const String headerAuthorizationBearer = 'Bearer';
+
+  /// the url for the cdn: https://cdn.ourjourneys.lukecreated.com
+  static const String cdnUrl = 'https://cdn.ourjourneys.lukecreated.com';
 }